dev-requirements.txt
discord.py
aio-pika
pytest
black
